# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Hosting API
  version: 1.0.0
  description: |
    Production-grade federated static site hosting network.
    Every node exposes this API. Authentication uses session cookies (browser)
    or Bearer API tokens (CLI / CI). Inter-node requests use Ed25519 signatures.
  contact:
    name: The No Hands Company
    url: https://github.com/The-No-Hands-company/Nexus-Hosting
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Local node API
...
paths:
...
  /sites/{id}/files/upload-url:
    post:
      operationId: getSiteFileUploadUrl
      tags: [deploy]
      summary: Get a presigned URL to upload a site file
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteFileUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteFileUploadUrlResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /sites/{id}/files:
    get:
      operationId: listSiteFiles
      tags: [deploy]
      summary: List all files registered for a site
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: List of site files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteFile"
    post:
      operationId: registerSiteFile
      tags: [deploy]
      summary: Register an uploaded file for a site
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSiteFileBody"
      responses:
        "201":
          description: File registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteFile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sites/{id}/deploy:
    post:
      operationId: deploySite
      tags: [deploy]
      summary: Atomically deploy all pending files and notify federation peers
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deployment activated with replication results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sites/{id}/deployments:
    get:
      operationId: listSiteDeployments
      tags: [deploy]
      summary: List deployment history for a site
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deployment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteDeployment"
  /sites/{id}/deployments/{depId}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deploy]
      summary: Roll back to a specific previous deployment version
      description: |
        Creates a new active deployment pointing to the files from the target
        deployment. The current active deployment is marked as rolled_back.
        History is fully preserved — rollback is a forward operation.
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: depId
          in: path
          required: true
          schema:
            type: integer
          description: ID of the deployment to roll back to
      responses:
        "200":
          description: Rollback successful — new deployment record created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SiteDeployment"
                  - type: object
                    properties:
                      rolledBackFrom:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Access Control ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Browser session cookie set by /login flow
    bearerToken:
      type: http
      scheme: bearer
      description: API token created via POST /tokens (prefix fh_)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
...
    SiteFile:
      type: object
      required: [id, siteId, filePath, objectPath, contentType, sizeBytes, createdAt]
      properties:
        id:
          type: integer
        siteId:
          type: integer
        deploymentId:
          type: ["integer", "null"]
        filePath:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
    SiteFileUploadUrlBody:
      type: object
      required: [filePath, contentType, size]
      properties:
        filePath:
          type: string
        contentType:
          type: string
        size:
          type: integer
    SiteFileUploadUrlResponse:
      type: object
      required: [uploadUrl, objectPath, filePath]
      properties:
        uploadUrl:
          type: string
        objectPath:
          type: string
        filePath:
          type: string
    RegisterSiteFileBody:
      type: object
      required: [filePath, objectPath, contentType, sizeBytes]
      properties:
        filePath:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
    SiteDeployment:
      type: object
      required: [id, siteId, version, status, fileCount, totalSizeMb, deployedAt]
      properties:
        id:
          type: integer
        siteId:
          type: integer
        version:
          type: integer
        status:
          type: string
          enum: [pending, active, failed, rolled_back]
        deployedBy:
          type: ["string", "null"]
        fileCount:
          type: integer
        totalSizeMb:
          type: number
        deployedAt:
          type: string
        createdAt:
          type: string
    DeployResult:
      allOf:
        - $ref: "#/components/schemas/SiteDeployment"
        - type: object
          properties:
            replication:
              type: object
              properties:
                peers:
                  type: integer
                synced:
                  type: integer
                results:
                  type: array
                  items:
                    type: object
                    properties:
                      node:
                        type: string
                      success:
                        type: boolean
                      error:
                        type: ["string", "null"]
...